backend/requirements.txt
google-generativeai
# openai (deprecated, switching to google-generativeai)
alpaca-py
yfinance
pandas
tabulate
requests                                                                                                                                        
transformers
python-dotenv
pymongo
matplotlib
jupyter
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
duckduckgo_search==8.0.1
fastapi==0.115.12
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
lxml==5.4.0
mcp==1.6.0
# openai==1.76.0
# openai-agents==0.0.13
primp==0.15.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
sse-starlette==2.3.3
starlette==0.46.2
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
